agent-runner/requirements.txt
redis==5.0.1
psycopg2-binary==2.9.9
requests==2.31.0
python-dotenv==1.0.0
openai==1.12.0
httpx==0.27.0
pydantic==2.5.2
aiohttp==3.9.1
